+++ CHANGES.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 - pythonProject: use latest pip-tools and remove pins for setuptools and pip 
+- setuptools compatibility: convert git describe output to pep440 compatible version
 
 ## 2023-01-30 / 1.11.0